requirements.txt
dj-database-url==0.4.1
dj-static==0.0.6
Django==1.10.4
python-decouple==3.0
static3==0.7.0
gunicorn==19.6.0
psycopg2==2.6.2
django-extensions==1.7.5
django-test-without-migrations==0.4
django-debug-toolbar
djangorestframework==3.5.3
django-cors-headers==1.3.1
django-mptt==0.8.6
-e git://github.com/darklow/django-suit.git@v2#egg=django-suit